+++ src/components/Popup/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Popup'
+export { default as PopupContent } from './PopupContent'